[SCISERVER] Revert 2 commits related to appLogPrintf -> Uart in MCU1_0
The fix for ADASVISION-5669 in the reverted commits here is causing issues
on some boards when vision_apps builds MCU1_0 image, but not other boards.
Reverting is allowing this to work on all tested boards.  Reverting for
now until this can be debugged further.

Revert "utils: sciserver: app: remove redefinition of variables for AM62A"
Remove reinclusion of header files and redeclaration of variables for AM62A.
This reverts commit 9fb3ddf870bbe98a486a5f0a10c18c083eba5178.

Revert "Using UART prints instead of appLogPrintf"
- Fixes ADASVISION-5669
This reverts commit 231e593328e425446c6d964ed051332f5c8caae7.