[MAKERULES]: Fix vision apps firmware softlinke for AM62A
AM62A SoC uses wkup R5 for running vision apps
along with Device Manager early in the bootup
cycle so there is no soft link needed in rootfs
for the same.

There is a separate MCU R5 actually and the
softlink for the same was getting deleted while
doing vision apps install.

Fix it by only allowing to delete c7x softlinks
that is the only one which need to be updated
by vision apps.

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>